# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SM Backend API
  version: 1.0.0
  description: API spec generated from current source controllers.
servers:
  - url: http://localhost:3000
    description: Local NestJS
  - url: https://be-sm.codekhongngu.workers.dev
    description: Cloudflare Worker
...
paths:
...
  /evaluations:
    post:
      tags: [Evaluations]
      summary: Create evaluation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
  /evaluations/{journalId}:
    patch:
      tags: [Evaluations]
      summary: Update evaluation by journal
      security:
        - bearerAuth: []
      parameters:
        - name: journalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /evaluations/{journalId}/awareness:
    patch:
      tags: [Evaluations]
      summary: Submit awareness form
      security:
        - bearerAuth: []
      parameters:
        - name: journalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /evaluations/{journalId}/standards:
    patch:
      tags: [Evaluations]
      summary: Submit standards form
      security:
        - bearerAuth: []
      parameters:
        - name: journalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
  /evaluations/pending/list:
    get:
      tags: [Evaluations]
      summary: Pending evaluations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /evaluations/analytics/weekly:
    get:
      tags: [Evaluations]
      summary: Weekly analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Weekly analytics
          content:
            application/json:
              schema:
                type: object
  /journals:
    get:
      tags: [Journals]
      summary: Get journals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Journal list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      tags: [Journals]
      summary: Create journal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
  /journals/{id}:
    get:
      tags: [Journals]
      summary: Get journal by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Journal
          content:
            application/json:
              schema:
                type: object
  /journals/eform-awareness:
    post:
      tags: [Journals]
      summary: Submit awareness eform
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                type: object
  /journals/eform-standards:
    post:
      tags: [Journals]
      summary: Submit standards eform
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                type: object
...
  /api/logs/submit:
    post:
      tags: [Behavior]
      summary: Submit behavior log by form type
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitLogRequest"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                type: object
  /api/manager/logs/pending:
    get:
      tags: [Behavior]
      summary: Get pending form-2 logs in current week
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  week:
                    $ref: "#/components/schemas/WeeklyConfig"
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/BehaviorChecklistLog"
  /api/manager/logs/evaluate/{id}:
    patch:
      tags: [Behavior]
      summary: Evaluate form-2 log
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateBehaviorLogRequest"
      responses:
        "200":
          description: Evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BehaviorChecklistLog"
  /api/reports/summary/weekly/{weekId}:
    get:
      tags: [Behavior]
      summary: Weekly behavior summary by manager evaluation
      security:
        - bearerAuth: []
      parameters:
        - name: weekId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Weekly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyBehaviorSummaryResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    SubmitLogRequest:
      type: object
      properties:
        formType:
          type: string
          enum: [FORM_1, FORM_2, FORM_3, FORM_4, FORM_5, FORM_8]
        logDate:
          type: string
          format: date
        avoidance:
          type: string
        selfLimit:
          type: string
        earlyStop:
          type: string
        blaming:
          type: string
        askedDeepQuestion:
          type: boolean
        fullConsultation:
          type: boolean
        followedThrough:
          type: boolean
        customerMetCount:
          type: integer
          minimum: 0
        employeeNotes:
          type: string
        negativeThought:
          type: string
        newMindset:
          type: string
        behaviorChange:
          type: string
        salesActivities:
          type: array
          items:
            type: object
            properties:
              customerName:
                type: string
              customerIssue:
                type: string
              consequence:
                type: string
              solutionOffered:
                type: string
              valueBasedPricing:
                type: string
              result:
                type: string
        differentAction:
          type: string
        customerImpact:
          type: string
        tomorrowLesson:
          type: string
        beliefTransformations:
          type: array
          items:
            type: object
            properties:
              situation:
                type: string
              oldBelief:
                type: string
              newChosenBelief:
                type: string
              newBehavior:
                type: string
              result:
                type: string
        situation:
          type: string
        oldBelief:
          type: string
        newChosenBelief:
          type: string
        newBehavior:
          type: string
        transformationResult:
          type: string
      required: [formType]
    EvaluateBehaviorLogRequest:
      type: object
      properties:
        mgrEvalDeepQ:
          type: boolean
        mgrEvalFullCons:
          type: boolean
        mgrEvalFollow:
          type: boolean
        managerFeedback:
          type: string
      required: [mgrEvalDeepQ, mgrEvalFullCons, mgrEvalFollow]
    WeeklyConfig:
      type: object
      properties:
        id:
          type: string
        weekName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required: [id, weekName, startDate, endDate]
    BehaviorChecklistLog:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        logDate:
          type: string
          format: date
        askedDeepQuestion:
          type: boolean
        fullConsultation:
          type: boolean
        followedThrough:
          type: boolean
        customerMetCount:
          type: integer
        employeeNotes:
          type: string
        managerId:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        mgrEvalDeepQ:
          type: boolean
          nullable: true
        mgrEvalFullCons:
          type: boolean
          nullable: true
        mgrEvalFollow:
          type: boolean
          nullable: true
        managerFeedback:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, logDate, status]
    WeeklyBehaviorSummaryResponse:
      type: object
      properties:
        week:
          $ref: "#/components/schemas/WeeklyConfig"
        totalDays:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              totalCustomerMet:
                type: string
              deepQuestionPct:
                type: string
              fullConsultationPct:
                type: string
              followedThroughPct:
                type: string